third_party/blink/web_tests/wpt_internal/app-history/navigate-event/respondWith-same-document-history-back.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
async_test(t => {
  // step_timeout because blink doesn't create history entries if onload has not
  // yet completed.
  window.onload = () => t.step_timeout(() => {
    let onnavigate_calls = 0;
    let onnavigatesuccess_calls = 0;
    appHistory.onnavigate = e => {
      onnavigate_calls++;
      e.respondWith(Promise.resolve());
    }
    appHistory.onnavigatesuccess = t.step_func(e => {
      onnavigatesuccess_calls++;
      if (onnavigatesuccess_calls == 2) {
        assert_equals(appHistory.entries().length, 2);
        assert_equals(appHistory.current.index, 0);
        assert_equals(onnavigate_calls, 2);
        t.done();
      }
    });

    appHistory.navigate("?foo").then(t.step_func(() => {
      assert_equals(appHistory.entries().length, 2);
      assert_equals(appHistory.current.index, 1);
      assert_equals(onnavigate_calls, 1);
      history.back();
    }));
  }, 0);
}, "AppHistoryNavigateEvent.respondWith can intercept same-document history.back()");
</script>
